move stuff inside lib folder
