+++ server/database/lib/general/query-types.ts
