+++ server/database/lib/general/serialized-types.ts
